+++ bin/actions/0220022-57.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+#: ◣ + SPACE - in terminal - Type "yes"
+
+focusApplication=\
+$(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
+
+if [[ "$focusApplication" =~ "gnome-terminal-" ]]; then
+  xdotool type --delay 0 "yes"
+  xdotool key Return
+fi