+++ bin/actions/0220022-78.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+#: ◣ + numeric [+] - Open xdeskmenu
+
+xdeskmenu -pp -o=nw